# packages/api-schema/openapi.json
{
  "components": {
    "schemas": {
...
      "Appeal": {
        "description": "A challenge to a suspension.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "default": "pending",
            "title": "Status",
            "type": "string"
          },
          "suspension_id": {
            "title": "Suspension Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "suspension_id"
        ],
        "title": "Appeal",
        "type": "object"
      },
      "Block": {
        "description": "Somebody the caller has blocked.",
        "properties": {
          "blocked_id": {
            "title": "Blocked Id",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "required": [
          "blocked_id"
        ],
        "title": "Block",
        "type": "object"
      },
...
      "CreateAppealRequest": {
        "description": "Challenge a suspension, with a written statement.",
        "properties": {
          "statement": {
            "maxLength": 2000,
            "minLength": 10,
            "title": "Statement",
            "type": "string"
          },
          "suspension_id": {
            "title": "Suspension Id",
            "type": "string"
          }
        },
        "required": [
          "suspension_id",
          "statement"
        ],
        "title": "CreateAppealRequest",
        "type": "object"
      },
      "CreateBlockRequest": {
        "description": "Block another student.\n\nBlocking hides content in both directions, so it also stops the\nblocked user seeing the blocker.",
        "properties": {
          "blocked_id": {
            "title": "Blocked Id",
            "type": "string"
          }
        },
        "required": [
          "blocked_id"
        ],
        "title": "CreateBlockRequest",
        "type": "object"
      },
      "CreateJoinRequest": {
        "description": "Ask the host for a seat, with an optional note.",
        "properties": {
          "message": {
            "anyOf": [
              {
                "maxLength": 280,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "title": "CreateJoinRequest",
        "type": "object"
      },
      "CreateReportRequest": {
        "description": "Flag a profile, table, or message.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "reason": {
            "maxLength": 100,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "subject_type": {
            "$ref": "#/components/schemas/SubjectType"
          }
        },
        "required": [
          "subject_type",
          "subject_id",
          "reason"
        ],
        "title": "CreateReportRequest",
        "type": "object"
      },
...
      "CreateTableRequest": {
        "description": "Host a new table. The host takes the first seat automatically.",
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 280,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "restaurant_id": {
            "title": "Restaurant Id",
            "type": "string"
          },
          "scheduled_for": {
            "format": "date-time",
            "title": "Scheduled For",
            "type": "string"
          },
          "seats_total": {
            "maximum": 12.0,
            "minimum": 2.0,
            "title": "Seats Total",
            "type": "integer"
          },
          "vibe_tag_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Vibe Tag Ids",
            "type": "array"
          }
        },
        "required": [
          "restaurant_id",
          "scheduled_for",
          "seats_total"
        ],
        "title": "CreateTableRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JoinRequest": {
        "description": "A student asking for a seat.\n\nAttributes:\n    match_score: Snapshot taken when the request was made, so the\n        host's view stays stable even after the scoring model\n        changes.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "match_score": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JoinRequestStatus",
            "default": "pending"
          },
          "table_id": {
            "title": "Table Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "table_id",
          "profile_id"
        ],
        "title": "JoinRequest",
        "type": "object"
      },
      "JoinRequestStatus": {
        "description": "Lifecycle of one student's request to join a table.",
        "enum": [
          "pending",
          "accepted",
          "declined",
          "withdrawn"
        ],
        "title": "JoinRequestStatus",
        "type": "string"
      },
...
      "Report": {
        "description": "Something a user flagged for moderation.",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "default": "open",
            "title": "Status",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "subject_type": {
            "$ref": "#/components/schemas/SubjectType"
          }
        },
        "required": [
          "id",
          "subject_type",
          "subject_id",
          "reason"
        ],
        "title": "Report",
        "type": "object"
      },
...
      "ResolveJoinRequest": {
        "description": "The host's decision on a pending request.",
        "properties": {
          "accept": {
            "title": "Accept",
            "type": "boolean"
          }
        },
        "required": [
          "accept"
        ],
        "title": "ResolveJoinRequest",
        "type": "object"
      },
      "Restaurant": {
        "description": "One restaurant in the catalogue.\n\nAttributes:\n    price_level_cents: Rough per-head cost in integer minor units.\n        Never a float \u2014 see the data conventions in the README.\n    distance_km: Populated only when the request supplied a\n        location.",
        "properties": {
          "address_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "cuisine_tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Cuisine Tags",
            "type": "array"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "price_level_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Level Cents"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "Restaurant",
        "type": "object"
      },
      "RestaurantPage": {
        "description": "A cursor-paginated page of restaurants.",
        "properties": {
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Restaurant"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "RestaurantPage",
        "type": "object"
      },
...
      "SubjectType": {
        "description": "What a report is about.",
        "enum": [
          "profile",
          "table",
          "message"
        ],
        "title": "SubjectType",
        "type": "string"
      },
...
      "Suspension": {
        "description": "A period during which an account cannot act.",
        "properties": {
          "ends_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reason",
          "starts_at"
        ],
        "title": "Suspension",
        "type": "object"
      },
...
      "TablePage": {
        "description": "A cursor-paginated page of tables.",
        "properties": {
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DiningTable"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "TablePage",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The only seam between the Expo app and Tablr's data. The app never talks to Postgres directly.",
    "title": "Tablr API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/appeals": {
      "post": {
        "description": "Challenge a suspension.",
        "operationId": "create_appeal_v1_appeals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppealRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appeal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Appeal",
        "tags": [
          "trust-safety"
        ]
      }
    },
    "/v1/blocks": {
      "get": {
        "description": "Return everybody the caller has blocked.",
        "operationId": "list_blocks_v1_blocks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Block"
                  },
                  "title": "Response List Blocks V1 Blocks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Blocks",
        "tags": [
          "trust-safety"
        ]
      },
      "post": {
        "description": "Block another student.",
        "operationId": "create_block_v1_blocks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Block",
        "tags": [
          "trust-safety"
        ]
      }
    },
    "/v1/blocks/{blocked_id}": {
      "delete": {
        "description": "Unblock another student.",
        "operationId": "remove_block_v1_blocks__blocked_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "blocked_id",
            "required": true,
            "schema": {
              "title": "Blocked Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Block",
        "tags": [
          "trust-safety"
        ]
      }
    },
    "/v1/join-requests/{request_id}": {
      "patch": {
        "description": "Accept or decline a pending request. Host only.",
        "operationId": "resolve_join_request_v1_join_requests__request_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveJoinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve Join Request",
        "tags": [
          "tables"
        ]
      }
    },
...
    "/v1/me/suspensions": {
      "get": {
        "description": "Return the caller's active suspensions.",
        "operationId": "list_my_suspensions_v1_me_suspensions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Suspension"
                  },
                  "title": "Response List My Suspensions V1 Me Suspensions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Suspensions",
        "tags": [
          "trust-safety"
        ]
      }
    },
    "/v1/me/tables": {
      "get": {
        "description": "Return tables the caller hosts or belongs to.",
        "operationId": "list_my_tables_v1_me_tables_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TablePage"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "My tables",
        "tags": [
          "tables"
        ]
      }
    },
...
    "/v1/reports": {
      "post": {
        "description": "Flag a profile, table, or message for moderation.",
        "operationId": "create_report_v1_reports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Report",
        "tags": [
          "trust-safety"
        ]
      }
    },
    "/v1/restaurants": {
      "get": {
        "description": "Return a page of restaurants matching the query.",
        "operationId": "list_restaurants_v1_restaurants_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "near_lat",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 90,
                  "minimum": -90,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Near Lat"
            }
          },
          {
            "in": "query",
            "name": "near_lng",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 180,
                  "minimum": -180,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Near Lng"
            }
          },
          {
            "in": "query",
            "name": "radius_km",
            "required": false,
            "schema": {
              "default": 5,
              "exclusiveMinimum": 0,
              "maximum": 50,
              "title": "Radius Km",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "cuisine",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cuisine"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Browse",
        "tags": [
          "restaurants"
        ]
      }
    },
    "/v1/restaurants/{restaurant_id}": {
      "get": {
        "description": "Return one restaurant by id.",
        "operationId": "get_restaurant_v1_restaurants__restaurant_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "restaurant_id",
            "required": true,
            "schema": {
              "title": "Restaurant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Restaurant"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Restaurant",
        "tags": [
          "restaurants"
        ]
      }
    },
    "/v1/tables": {
      "get": {
        "description": "Return open tables the caller could ask to join.",
        "operationId": "list_open_tables_v1_tables_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TablePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Open tables",
        "tags": [
          "tables"
        ]
      },
      "post": {
        "description": "Host a new table.",
        "operationId": "create_table_v1_tables_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiningTable"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Table",
        "tags": [
          "tables"
        ]
      }
    },
    "/v1/tables/{table_id}": {
      "get": {
        "description": "Return one table the caller is allowed to see.",
        "operationId": "get_table_v1_tables__table_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiningTable"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Table",
        "tags": [
          "tables"
        ]
      }
    },
    "/v1/tables/{table_id}/join-requests": {
      "post": {
        "description": "Ask the host for a seat at this table.",
        "operationId": "request_to_join_v1_tables__table_id__join_requests_post",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJoinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request To Join",
        "tags": [
          "tables"
        ]
      }
    },
...
  }
}